ADM-990 [frontend]: try to fix fossa issue
