+++ frontend/__tests__/containers/ConfigStep/SourceControl.test.tsx
@@ -13,6 +13,7 @@ import {
 import { sourceControlDefaultValues } from '@src/containers/ConfigStep/Form/useDefaultValues';
 import { sourceControlClient } from '@src/clients/sourceControl/SourceControlClient';
 import { AxiosRequestErrorCode, SourceControlTypes } from '@src/constants/resources';
+import { updateShouldGetPipelineConfig } from '@src/context/Metrics/metricsSlice';
 import { sourceControlSchema } from '@src/containers/ConfigStep/Form/schema';
 import { SourceControl } from '@src/containers/ConfigStep/SourceControl';
 import { render, screen, act, waitFor } from '@testing-library/react';
@@ -63,15 +64,13 @@ describe('SourceControl', () => {
     sourceControlClient.verifyToken = originalVerifyToken;
   });
 
-  const onReset = jest.fn();
-  const onSetResetFields = jest.fn();
   store = setupStore();
   const setup = () => {
     store = setupStore();
     return render(
       <Provider store={store}>
         <FormProvider schema={sourceControlSchema} defaultValues={sourceControlDefaultValues}>
-          <SourceControl onReset={onReset} onSetResetFields={onSetResetFields} />
+          <SourceControl />
         </FormProvider>
       </Provider>,
     );
@@ -91,21 +90,6 @@ describe('SourceControl', () => {
     const sourceControlType = screen.getByText(SourceControlTypes.GitHub);
 
     expect(sourceControlType).toBeInTheDocument();
-  });
-
-  it('should run the reset and setResetField func when click reset button', async () => {
-    setup();
-    await fillSourceControlFieldsInformation();
-
-    await userEvent.click(screen.getByRole('button', { name: VERIFY }));
-
-    await waitFor(async () => {
-      expect(screen.getByRole('button', { name: RESET })).toBeTruthy();
-      await userEvent.click(screen.getByRole('button', { name: RESET }));
-    });
-
-    expect(onReset).toHaveBeenCalledTimes(1);
-    expect(onSetResetFields).toHaveBeenCalledTimes(1);
   });
 
   it('should hidden timeout alert when the error type of api call becomes other', async () => {
@@ -154,6 +138,19 @@ describe('SourceControl', () => {
     await waitFor(() => {
       expect(screen.getByText(VERIFIED)).toBeTruthy();
     });
+  });
+
+  it('should reload pipeline config when reset fields', async () => {
+    setup();
+    await fillSourceControlFieldsInformation();
+
+    await userEvent.click(screen.getByText(VERIFY));
+
+    await userEvent.click(screen.getByRole('button', { name: RESET }));
+
+    await fillSourceControlFieldsInformation();
+
+    expect(updateShouldGetPipelineConfig).toHaveBeenCalledWith(true);
   });
 
   it('should show error message and error style when token is empty', async () => {
